extracted attributes into separate files; extracted macOS specific code to separate folder and update podspec
