+++ Sources/OzStringKit/MacOS/MacOSTypes.swift
@@ -1,0 +1,14 @@
+//
+//  MacOSTypes.swift
+//  OzStringKit
+//
+//  Created by Konstantin Oznobikhin on Mar 21, 2020.
+//  Copyright © 2020 Konstantin Oznobikhin. All rights reserved.
+//
+
+#if os(macOS)
+    import AppKit
+    public typealias OzImage = NSImage
+    public typealias OzColor = NSColor
+    public typealias OzFont = NSFont
+#endif
