+++ Sources/OzStringKit/Types.swift
@@ -1,0 +1,16 @@
+//
+//  Types.swift
+//  OzStringKit
+//
+//  Created by Konstantin Oznobikhin on Mar 20, 2020.
+//  Copyright © 2020 Konstantin Oznobikhin. All rights reserved.
+//
+
+import Foundation
+
+#if !os(macOS)
+    import UIKit
+    public typealias OzImage = UIImage
+    public typealias OzColor = UIColor
+    public typealias OzFont = UIFont
+#endif
